pyproject.toml
[project]
name = "agarmmo"
version = "0.1.0"
description = "A multiplayer game inspired by Agar.io, built with Python, Pygame, and sockets."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.5",
    "pyyaml>=6.0.2",
]
